feat: implement home screen with actions, movement history, and synchronization logic.

+++ app/src/main/java/com/michambita/domain/model/MovimientoItem.kt
@@ -12,5 +12,5 @@ data class MovimientoItem(
 fun MovimientoItemModel.toDomain() = MovimientoItem(
     productoId,
     cantidad,
-    precioTotal
+    precioTotal.toBigDecimal()
 )
